QL_tour/requirements.txt
django>=4.2
pillow
psycopg2-binary
djangorestframework
djangorestframework-gis
leafmap
GDAL==3.7.0
django-geojson